Day 10!
Lotta crap on the subreddit for using recursion, but recursion with
memoization finishes in 0.033s user time in this solution on my MBP, so
really not bad. I did get bit by the numbers being larger than INT_MAX
and then missed one place for int64_t (in the memo dict). Oops :(