Added a linear solution for Day 10.
The linear solution is a cute observation of how the memoized recursive
function works in practice, which is that it's basically that it's
walking the list backwards. You know that the last element has exactly
one path to the final joltage value, and then you walk backwards an
adapter at a time and add up the paths it can reach. When you're done,
you've generated the equivalent of the memo and can just add up the
paths from adapters of values 1-3 (some of which may not exist).

If you're memoizing the recursive solution, it's not terribly different
in terms of runtime (you only visit each value once), but you save the
stack frame each time, so it's better.