Day 13, extremely meh.
This is exactly the type of problem that we're not supposed to ask at
work anymore, it's a total "gotcha" problem. The brute force solution
would take days (at best), even if you do cute things like only check
things in the congruence modulo space for the largest bus id (e.g. my
answer was 266204454441577 and my largest bus id is 643, so that's a
lot of checks for the other bus_id congruences).

But if you get the "gotcha" that its the Chinese remainder theorem
because they are all coprimes(?!) and you remember (?!?!) the algorithm
for solving it, it's...easy! But that's an obnoxious way to solve
problems, and this is everything I was hoping advent of code was not.

What a terrible puzzle :(