Day 14, ugh.
Spent a lot of time debugging part 2 only to finally realize that I was
building digit-setting masks as:

int64_t on = address | (1 << digit);
int64_t off = address & ~(1 << digit);

Can you spot the bug?

"1 << digit" is an integer, forcing the whole thing to int32_t and then
extending it to on/off as int64_t. It needs to be:
1ll << digit

Or similar. Bleh.