Day 15.
This feels like an interview question, particularly since the Part 1
answer can be done the "dumb" way (store the numbers in a vector, walk
backwards to find distance, so: O(n^2)) but Part 2 needs to be redone
in an more optimal way (store a number->last seen index map and
calculate offset with a map lookup, so: O(~n)).

This is also one of those cases where it's easy to end up off by one in
various places. Smarter people probably reason it out but I tend to just
move things around until it's correct (e.g. my index is off by one so I
look for 2019 instead of 2020, etc.).