Day 18!
This was a kinda fun one, also I cheated a bit.

I was trying to remember how to do expressions and was at a loss, so I
did two things:
1) I checked reddit to see which way people were going, which is where I
   saw a meme that Part 2 was going to change operator precedence, so I
   realized I should make the solution generic in that way (I originally
   was just going to blindly do it).
2) I remembered doing something like this in Scheme, where the trick was
   just to convert infix (which requires precedence rules) to
   S-expressions (which has no precedence). Then I remembered you could
   do the same with RPN, but I couldn't remember the algorithm. So I
   looked it up on Wikipedia (Shunting-yard) and implemented it from the
   psuedocode there.

Shunting-yard is actually really simple once you know it, you just need
the rules for when/how you move things from the operator queue to the
output queue. Once you have that, you plug in precedence for each part
and then write a simple evaluator for RPN, which looks a bit like
Shunting-yard but just uses a single stack where you push numbers and
then, when you encounter an operator, use it combine the last two
numbers. When you're done, you should be left with a stack with a single
number, which is your answer.